requirements.txt
streamlit
google-generativeai
python-dotenv
requests
beautifulsoup4
PyPDF2
python-dateutil
